Added login step defs to main file
